Test for Packages folder before extracting content

+++ NugetDbTools/NuGetSharedPacker/Scripts/Get-SolutionContent.ps1
@@ -41,7 +41,7 @@ function Get-SolutionContent {
 
 	del $packageContentFolder -Include '*' -Recurse -Force
 
-	if (ls "$packageFolder\**\$contentFolder" -Recurse) {
+	if ((Test-Path $packageFolder) -and (ls "$packageFolder\**\$contentFolder" -Recurse)) {
 		if (Test-Path $solutionContentFolder) {
 			rmdir $solutionContentFolder* -Recurse -Force
 		}
